Do not emit trackpad events in scroll mode
